fix(APISIMP-PREDICATES-PAGECAP): prevent CWE-190 int overflow in skip arithmetic
Use long skip = (long) page * pageSize in ShapesPredicatesRest to
prevent integer overflow at large page/pageSize values (CWE-190).
Update PredicateVocabularyRepository.findAll and findBySubstrate to
accept long skip; change ps.setInt(2/3, skip) to ps.setLong for OFFSET.

+++ backend/src/main/java/de/dlr/shepard/v2/shapes/resources/ShapesPredicatesRest.java
@@ -113,7 +113,7 @@ public class ShapesPredicatesRest {
     @Parameter(description = "Zero-based page index. Default 0.", required = false)
     @QueryParam("page") @DefaultValue("0") @PositiveOrZero int page
   ) {
-    int skip = page * pageSize;
+    long skip = (long) page * pageSize;
     long total;
     List<PredicateVocabularyEntryIO> items;
     if (substrate != null && !substrate.isBlank()) {
